test_requirements.txt
-r requirements.txt
coverage
coveralls
mock>=1.0.1
nose>=1.3.0
django-nose>=1.2
flake8>=2.1.0
tox>=1.8.0
ipdb

# Additional test requirements go here
factory_boy>=2.5
